+++ frontend/src/components/seller/signup/storeInfo/RegisterDesc.tsx
