+++ frontend/src/components/seller/signup/storeInfo/RegisterName.tsx
