+++ frontend/src/components/seller/signup/storeInfo/RegisterNum.tsx
