+++ frontend/src/components/seller/signup/storeInfo/index.ts
